+++ ReleaseNotes.md
@@ -3,3 +3,5 @@
 **Bug Fixes**
 
 **Technical Debt**
+
+- [Updating Automapper to latest version 3.2.1](https://trello.com/c/u5CNdGRz/139-update-automapper-to-latest-across-applications-and-use-nuget)